restores UserListener
